Change return type to int instead of enum
Enum is a better solution normally because it doesn't allow faulty
values to be passed (ones that aren't options) but in this case we want
to also be able to pass characters (which are sort of like the defined
values in enum as the number prepresents a letter. This is just less
code to do. To avoid collisions of states like NUMBER or COS we start
the enum of these values at >=256 (which is greater than the value of a
char in ASCII even with extended character set).